Refactor PlayerList into common components folder.
