+++ src/components/common/PlayerList/index.ts
@@ -1,0 +1,1 @@
+export { default as PlayerList } from './PlayerList';
