+++ src/components/pages/GameContainer/Game/index.ts
@@ -1,8 +1,8 @@
-import Lobby from './Lobby';
-import Pregame from './Pregame';
-import PlayerList from './PlayerList';
-import Writing from './Writing';
+import PlayerList from '../../../common/PlayerList/PlayerList';
 import Guessing from './Guessing';
+import Lobby from './Lobby';
 import Postgame from './Postgame';
+import Pregame from './Pregame';
+import Writing from './Writing';
 
 export { Lobby, Pregame, PlayerList, Writing, Guessing, Postgame };
